requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.12.5
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Segurança
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.9

# Cache
redis[hiredis]>=5.0.0

# Storage - Supabase Storage (S3-compatible)
boto3>=1.35.0

# Rate limiting
slowapi>=0.1.9

# HTTP client
httpx>=0.27.0

# Observabilidade
structlog>=24.0.0
